\name{StressStrength-package}
\alias{StressStrength-package}
\alias{StressStrength}
\docType{package}
\title{
Computation and Sample Estimation of Reliability of Stress-Strength Models
}
\description{
Reliability of (normal) stress-strength models and for building two-side or one-side confidence intervals according to different approximate procedures.
}
\details{
\tabular{ll}{
Package: \tab StressStrength\cr
Type: \tab Package\cr
Version: \tab 1.0.2\cr
Date: \tab 2016-04-29\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Alessandro Barbiero, Riccardo Inchingolo

Maintainer: Alessandro Barbiero <alessandro.barbiero@unimi.it>
}
\references{
Kotz S, Lumelskii Y, Pensky M (2003) The stress-strength model and its generalizations: theory
and applications. World Scientific, Singapore

Guo H, Krishnamoorthy K (2004) New approximate inferential methods for the reliability parameter
in a stress-strength model: The normal case. Commun Stat Theory Methods 33:1715-1731}
\keyword{ package }
